queue_files/requirements.txt
dropbox>=11.36.0
google-cloud-secret-manager>=2.16.1
google-cloud-pubsub>=2.17.1
requests>=2.31.0
mysql-connector-python>=8.0.33
